Update client.py to deal with no trainServices being returned by Darwin
The process was failing where there we no trains scheduled in the next 2 hours - in that instance, there is no "trainServices" object returned by Darwin and so the script failed to return anything to Home Assistant. 

This is amended by checking whether the raw batch results contain the "trainServices" object and then dealing with the situation accordingly when processing the raw data for Home Assistant.